sab: don't modify contact list selection when context menu popups
If we don't return TRUE, the selection will be changed to the item where the
user is right clicking. But we don't want that, if more than one row is
selected we want to keep these rows selected when right clicking to show up
the context menu (popup).

According to the documentation:
"Return TRUE to stop other handlers from being invoked for the event."
https://docs.gtk.org/gtk3/signal.Widget.button-press-event.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>