sab: close the contact editor window when the user presses 'Esc' key
From Microsoft UX Guidelines:

"Don't assign access keys to OK and Cancel buttons, because Enter is the
access key for the default button (which is usually the OK button),
and Esc is the access key for Cancel. Doing so makes the other
access keys easier to assign."
https://learn.microsoft.com/en-us/windows/win32/uxguide/ctrl-command-buttons#labels

"OK, Cancel, and Close buttons. Enter and Esc are used for their access keys."
https://learn.microsoft.com/en-us/windows/win32/uxguide/inter-keyboard#dialog-box-access-keys

"Spacebar, Enter, and Esc keys. The spacebar activates the control with input focus,
whereas the Enter key activates the default button. Pressing the Esc key cancels or closes the window."
https://learn.microsoft.com/en-us/windows/win32/uxguide/inter-keyboard#elements-of-keyboard-navigation

Signed-off-by: Thiago Farina <tfransosi@gmail.com>