sab: remove option to toggle the visibility of the menubar
Reverts commit 2620ba326a635490c39e004bfe6545077c7c7ba6 ("sab: add code to toggle the visibility of the menubar").

It doesn't seem a good option to provide to the users after all. They might
not remember the shortcut accelerator to bring it back.

Also, most of the programs in Windows 98 SE do not provide this option either.

So, let's remove it for now.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>