sab: add a Desktop Entry file
It is a configuration file that describes how a program is to be launched,
how it should appear in menus, etc.

That way, when I open the Applications menu in XFCE, an entry called 'Sab'
shows up under "Accessories" menu and when I click on it, it launches the gtk
application.

Tested with the following command lines:

$ sudo ninja -C build install

The specification for this Desktop file is at https://specifications.freedesktop.org/desktop-entry-spec/latest/

Most of the CMake code came from https://github.com/GNOME/evolution/blob/master/data/CMakeLists.txt

Signed-off-by: Thiago Farina <tfransosi@gmail.com>