sab: stop relying on gtk_tree_view_get_path_at_pos to get a valid GtkTreeIter
We were using gtk_tree_view_get_path_at_pos in order to get a GtkTreePath, and
then use it to get a valid GtkTreeIter for the selected row. As need to get
the contact pointer associated with it in order to show the properties dialog.

But that complicates the things, as it is hard to use this function
gtk_tree_view_get_path_at_pos in other contexts outside the event callbacks.

So, instead of relying on it, stop using the 'path' output and let's use
gtk_tree_selection_get_selected_rows to do the same. It is already used by other
code paths. It will give us more flexibility and allow us to move in a direction
where we can unify these code paths.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>