sab: break the return type from function names in window.c
We are trying to achieve a consistent code style here.

According to GNU Coding Standards:
"
5.1 Formatting Your Source Code
    It is also important for function definitions to start the name
    of the function in column one. This helps people to search for
    function definitions, and may also help certain tools recognize them.
"
https://www.gnu.org/prep/standards/html_node/Formatting.html

Function Declarations:
    Placing the name at the start of the line makes it easier to find functions
in a large codebase with e.g.:

    git grep '^my_function'

https://git.sr.ht/~sircmpwn/cstyle#function-declarations

So the main reason to put function names in column one is to make it easier
to search for them in large codebases.

It is --break-return-type in AStyle.
https://astyle.sourceforge.net/astyle.html#_attach-return-type

Signed-off-by: Thiago Farina <tfransosi@gmail.com>