sab: have the full cyle of DELETE prepared statement object inside '_db_delete_contact' function
The SQLite documentation says:

"
Prepared Statement Object
An instance of this object represents a single SQL statement that has been compiled into binary form and is ready to be evaluated.

Think of each SQL statement as a separate computer program. The original SQL text is source code. A prepared statement object is the compiled object code. All SQL must be converted into a prepared statement before it can be run.

The life-cycle of a prepared statement object usually goes like this:

   1. Create the prepared statement object using sqlite3_prepare_v2().
   2. Bind values to parameters using the sqlite3_bind_*() interfaces.
   3. Run the SQL by calling sqlite3_step() one or more times.
   4. Reset the prepared statement using sqlite3_reset() then go back to step 2. Do this zero or more times.
   5. Destroy the object using sqlite3_finalize().
"

https://www.sqlite.org/c3ref/stmt.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>