WunderPeople: use WINAPI calling convention instead of APIENTRY
APIENTRY is just an alias type declared as follows in WinDef.h:

WINAPI is a calling convention that defines how a function, in this case,
WinMain, receives parameters from the caller.

"The WinMain function is the same as wWinMain, except the command-line arguments are passed as an ANSI string."

WINAPI is the way recommned by Microsoft in its documentation, so let's just
use it.

Ref:
https://learn.microsoft.com/en-us/windows/win32/learnwin32/winmain--the-application-entry-point
https://learn.microsoft.com/en-us/windows/win32/winprog/windows-data-types

Signed-off-by: Thiago Farina <tfransosi@gmail.com>