sab: append 'cmd' suffix to 'sab' executable name
Change the name of the command line 'sab' executable to 'sabcmd' in order to
better identify which tool belongs to which environment. And to simplify the
name of the executable for the GNU/Linux GTK port.

That will help simplify things and be more consistent and will help with the Windows
port.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>