sab: remove the '_gui' suffix from the gtk executable name
This change makes it easier for the users when they are typing the
name of executable in the terminal. No need to press 'Shift' then
underscore + typing 'gui'.

The name is now much cleaner and easier to type. This will help with the
Windows port, which would have a 'sab.exe' as a counterpart.

So now, 'sabcmd' is the name of command line tool for the terminal and
'sab' is the name of GTK application for the graphical environment.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>