rename 'sab' project to 'tribe'
We chose another name for the address book application because we wanted
to have a more catchy word. A name that sounded nice, simple and elegant.

We also wanted a name that was short and easy to pronounce.

We chose 'tribe' or 'Tribe' because it is a synonym of 'people'.

https://www.merriam-webster.com/thesaurus/people

It will help with the Windows port. Currently the Windows port
has a name of 'WunderPeople', but we wanted something simpler, just
a single word, that could be easier to type, as tribe.exe is much
easier to type than WunderPeople.exe or wunderpeople.exe.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>