rename 'WunderPeople' project to 'Tribe'
We chose another name for the address book application because we wanted
to have a more catchy word. A name that sounded nice, simple and elegant.

We also wanted a name that was short and easy to pronounce. Although,
'WunderPeople' was fine, it had two words.

We chose 'tribe' because it is a synonym of 'people'.

https://www.merriam-webster.com/thesaurus/people

We wanted something simpler, just a single word, that could be
easier to type. For example, tribe.exe is much easier to type
than WunderPeople.exe or wunderpeople.exe.

This consolidates the names of 'sab' and 'WunderPeople' address book projects,
under a single name that is easy to pronounce and remember.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>