Tribe: go to the exit point if we are out of memory
If we failed to allocate memory, the user might be without memory
available, and we need to exit.

We might give a polite message and abort the application. But we are
not at that point yet.

TODO: Now we need to check the return value of 'load_contacts' function,
and do something about it, if it returned an error.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>