Tribe: add the script file used to create the Setup program
This change adds the script file for Inno Setup that controls
every aspect of the installation. It specifies which files are
to be installed and where, and so on.

The script was compiled with Inno Setup version 5.6.1, downloaded
from https://files.jrsoftware.org/is/5/?C=M;O=D.

The version 5.x is the major and last version to support Windows XP,
which is the system were building.

https://jrsoftware.org/isdlold.php
https://jrsoftware.org/files/is5-whatsnew.htm

This script was created with the help of the tutorial in https://youtube.com/watch?v=wP9HcD4oPDW ("Creating Custom Installers For Your Programs Using Inno Setup - Tutorial").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>