Tribe: add a license
Add a COPYING file to adhere to The GNU Coding Standards, as it indicates
in '7.3 Making Releases' that if a GNU GPL license is used, then it should
be in a file called COPYING.

https://www.gnu.org/prep/standards/html_node/Releases.html#index-version-numbers_002c-for-releases

Also add a LICENSE.txt file, which is what is recommended by Github. The '.txt'
extension is to make it easier to open on Windows.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>