start the reorganization of folder structure
Now that Tribe has been moved from https://github.com/tfarina/maple3,
into its own repository, we can start changing where the files are located
and better organize them, in a layout that reflects a standalone open source
project.

The goal is to make it easy for anyone to clone the project, make changes,
build, and do pretty much everything they want with it, without checking out
a bunch of unrelated stuff.

This way we can concentrate in the Windows port.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>