allow the user to delete an item using the menu item 'Delete'
Otherwise, when the user clicked this menu item, nothing would happen,
and that would be very frustrating for the user. He wouldn't know
what was wrong or why nothing happended, because it was essentially doing
nothing as it wasn't hooked up yet.

Now this menu item is functional and will do the action it is supposed to do.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>