add the bare minimium files necessary to create the .chm file
This change adds all the files necessary to create the .chm file for the
Microsoft HTML help system.

For now, it just adds the files and the compiled help manual (.chm) file.
Later on, we will copy the .chm into the installation directory to make it
available for the application to open it from the menu.

These files were created with the help of the tutorial "How to create .chm
files" at https://www.youtube.com/watch?v=BxVm_Edaus8 from PineHack.

The htmlhelp.exe was downloaded from https://www.helpandmanual.com/download/htmlhelp.exe and it installed Microsoft HTML Help 1.4.

Ref:
https://learn.microsoft.com/en-us/previous-versions/windows/desktop/htmlhelp/microsoft-html-help-1-4-sdk

Signed-off-by: Thiago Farina <tfransosi@gmail.com>