calculate the number of toolbar buttons dynamically
Let's not hard code the number of toolbar buttons, because we might add or
remove them in the future. If we have it as a constant, then we have to
remember to change it and that can be error prone and requires more changes
than necessary.

By calculating the array size to get the number of toolbar items we avoid that
issue and at the same time we make the code more resilient and reusable.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>