arraysize.h: remove hyphen from "statically-allocated"
It doesn't need the hyphen, the correct in English is without.

From grammarly "EHow to Use a Hyphen Correctly":

Don’t use a hyphen when you are combining an adverb and a participle.

Incorrect: The room was like a heavily-decorated chocolate box.

Correct: The room was like a heavily decorated chocolate box.

https://www.grammarly.com/blog/hyphen/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>