set toolbar button strings using TB_ADDSTRING message
The iString member of TBBUTTON can either a zero-based index
of the button string, or a pointer to a string buffer that
contains text for the button. Hence we have had to change the
tbButtons array.

We changed how strings for toolbar buttons are loaded and read, because
in the old way, it would be hard or even impossible to read them from
resource.

Changing to TB_ADDSTRING, makes the code more flexible and allows us
to move the string into a resource in the future if we want.

Ref:
https://learn.microsoft.com/en-us/windows/win32/api/commctrl/ns-commctrl-tbbutton
https://learn.microsoft.com/en-us/windows/win32/controls/create-toolbars

Signed-off-by: Thiago Farina <tfransosi@gmail.com>