populate contact's linked list from statically allocated array
We are trying to separate the data that cames from database to
the data presented to the user in the list view.

The data structure that represents an entry in the contacts table
could have dozens of properties, but when presenting to the user,
the list view would have a limited number of columns, say three or
four.

So we need a separate structure for each list view item and another
structure for each entry in the contacts table.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>