build the array of contacts in a separate function
We want to isolate the loading of contacts from the sqlite database,
to the contacts loaded in the list view.

Right now we are cheating and using a statically allocated array,
and passing that to a dynamic allocated array. But that will change
when we start loading the contacts directly from the database, the
static array will go away.

The goal is to have a clear separation and to store the contacts from
the database in a array while the contacts in the list view in a linked
list. Which is pretty much what we have achieved right now.

We are not loading from the database yet, because we are not able to
insert data there for now. But we are getting closer to achieve that as
well.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>