update menu and toolbar button states of 'Properties'
When the user clicks on an item in the list view, update the state of menu item
and the toolbar button accordingly.

Also handle the cases when the user clicks in the menu item or in the toolbar
button. For now, when that happens a "Not implemented yet" message is
displayed.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>