disable the annoying edit in place feature
The list view has this feature called LVS_EDITLABELS, that allows the user
to edit the item text of the first column by clicking on the row once,
sometimes twice. But that is annoying, since we don't support and want the user
to be changing the value of 'First Name' directly in the text row.

Finally we were able to disable it.

Ref:
https://learn.microsoft.com/en-us/windows/win32/controls/list-view-window-styles

Signed-off-by: Thiago Farina <tfransosi@gmail.com>