tribe: add the contact to the database first
Then check if something went wrong. If everything is ok, then add the contact
to the linked list.

I think if we do this way, in the future, we will be able to move the linked
list code out of the database code, and make the database code more stateless
leaving to the ui or client code to maintain that linked list.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>