change the icon used as app icon
This icon appears on the left side in the title bar along with the
name of the application. It's also used as the desktop shortcut.

Unfortunately, I didn't draw it myself. I took it from
https://www.iconfu.com/main#/icons/5/1/65/03b8e5-0-0, the material icons
free to use.

Finally, I was able to change the barcode icon to a new one. The old icon
came with the project template that Visual Studio provides when creating
a new application.

The change may look simple, but it was one of the hardest changes to
make in Visual Studio.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>