change the EmbedManifest setting to "false"
Otherwise, it does not do the right thing, it doesn't copy
tribe.exe.manifest into the Debug or Release directory, and then
the app looks crap. With EmbedManifest set to false, it copies
the file and the app looks good again.

I don't know why embedding does not work on Windows XP.

Ref: https://learn.microsoft.com/en-us/windows/win32/controls/cookbook-overview

Signed-off-by: Thiago Farina <tfransosi@gmail.com>