add a macro to replace MAX_LOADSTRING
MAX_LOADSTRING is a macro that came with the original Visual Studio
template for a win32 app. The name is not that good and has an arbitrary
size of 100.

To replace it, we are introducing another macro, called
MAX_STRING_RES_LENGTH, to give clear information to readers and writers
where and how it should be used. It should make the code easier to read
years later from now.

I chose the 255 size intentionally to keep the data storage at a single
byte which allow us to have 255 characters. If it was 256 or greater, two
bytes would be needed.
Ref:
https://stackoverflow.com/questions/2340649/why-historically-do-people-use-255-not-256-for-database-field-magnitudes

NOTE: This is the first change I'm committing and pushing directly from
the virtual machine running Windows XP.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>