check if there is an item selected when deleting
Since we are not calling UpdateUI() yet when the user presses the Delete
button on toolbar or menu, we are leaving the toolbar in a wrong state,
with buttons enabled when they shouldn't. So, if the user deletes
everything, we ended up in a situation where the Delete button is enabled
and the user can click on it, and that shows a message box asking if the
user wants to delete the selected item(s), which is wrong, because there
is nothing selected to delete to begin with.

That is why we have to call UpdateUI() in the end of delete code path, to
make sure the toolbar buttons are in the right state. Will do that in a
next patch.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>