call UpdateUI() after deleting an item
This is to make sure the toolbar goes into the right state updating the
buttons, enabling or disabling them, same for the menu.

If we don't call it, what happens is that after deleting the last item or
just everything, the Delete button stay enabled because it was not
updated. Calling UpdateUI() fixes that. See the previous commit for more
information.

We are working on setting the status bar text and while working on that
code, it made me realize this UI state issue with toolbar buttons and
menu.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>