for our confirmation dialog use a warning icon
By Microsoft's UX Guidelines definition:

An action that has some risk and can't be easily undone, which is the case
when the user deletes an item, is a risky action confirmation and because
they have a risk, they usually have a warning icon.

It also says that if the user is about to perform a risky action, a
confirmation is appropriate if the action cannot be easily undone.

Ref: https://learn.microsoft.com/en-us/windows/win32/uxguide/mess-confirm

Signed-off-by: Thiago Farina <tfransosi@gmail.com>