put focus in the "No" button in our confirmation dialog
From Microsoft's UX Guidelines:

According to them the default response for a confirmation should be based on
its design pattern and for a risky action confirmation the default
response is "Don't proceed (or the safe choice)".

So, to match this design pattern, let's set the default response for this
dialog in the "No" button.

Ref: https://learn.microsoft.com/en-us/windows/win32/uxguide/mess-confirm

Signed-off-by: Thiago Farina <tfransosi@gmail.com>