fix the inclusion of tribe.rcv in tribe.rc file
To make rc stop deleting it, I had to do the following:

1. Right-click the tribe.rc file in Resource View and select Resource
Includes.

2. In the read-only symbol directories, add the #include directive.

This is not the final solution though, because it still removes the usage
of APP_VERSION_STR macro from FileVersion and ProductionVersion. I don't
know the answer for why does it blow away the macro from them yet, so this
is only a partial solution while I keep investigating how to fully fix
this annoying issue.

Ref:
https://learn.microsoft.com/en-us/cpp/windows/how-to-include-resources-at-compile-time?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>