avoid having to do manual actions like copying files around (cp)
The file tribe.chm is created by create_chm.bat batch file in help/chm
directory and it outputs the generated file to that directory not to
src/Release. If we keep src/Release then we have to manually copy the file
there everytime, which is annoying, time consuming and we can forget from
time to time.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>