change the output directory of setup file
Output the setup file in the same directory as the Inno Setup script file.
It is easier to work this way, because I can open the directory in Windows
Explorer, compile the script and just see the file there next to
TribeX86.iss. If it goes to ../../Output, then I have to change
directories, which is annoying and time consuming.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>