move version resource block out of the main resource script file
I realized that in order to stop Resource Editor from throwing away my
macros and stop it from messing with my changes to version resource
strings, I had to move it out of tribe.rc file. And I figured that when
fixing the inclusion of tribe.rcv because when reading the documentation I
saw it says "shouldn't be modified" in Read-only symbol directives and
that was what made be realize what I was missing to fix this annoying
issue.

I didn't use APP_STR_VERSION in this change, but I will probably use it in
a next change.

Ref:
https://learn.microsoft.com/en-us/cpp/windows/how-to-include-resources-at-compile-time?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>