don't disable the Welcome Page in innosetup installer
They changed this setting in 5.5.7, setting it to yes, saying it was
because of a Microsoft desktop guideline.

I found this pretty annoying, I like the welcome page and I prefer to see
it first. I think for the users it is also better to welcome them to the
setup wizard. With this setting it nows shows the nice "Welcome to the
Tribe Setup Wizard" message.

Ref: https://jrsoftware.org/files/is5.5-whatsnew.htm#5.5.7

Signed-off-by: Thiago Farina <tfransosi@gmail.com>