rename Installer/InnoSetup to installer/innosetup
The goal is to have all directory names in lowercase. All other
directories were already in lowercase (e.g., help, chm, src). This is to
make file searches easier to do on UNIX-like operating systems like
GNU/Linux which are case sensitive (so you don't have to press Shift all
the time).

To make this change I had to change ignorecase config on Windows:

$ git config core.ignorecase false

And them issue the following commands:

$ git mv Installer/InnoSetup tempname
$ git mv Installer/tempname Installer/innosetup

$ git mv Installer tempname
$ git mv tempname installer

Ref:
https://stackoverflow.com/questions/11183788/in-a-git-repository-how-to-properly-rename-a-directory#11183844
https://developers.googe.com/style/filenames

Signed-off-by: Thiago Farina <tfransosi@gmail.com>