rename Tribe.iss to lowercase tribe.iss
The goal is to try to have almost all file names in lowercase. This is to
make file searches easier to do on UNIX-like operating systems like
GNU/Linux which are case sensitive (so you don't have to press Shift all
the time).

To make this change I had to change ignorecase config previously on
Windows with the following command:

$ git config core.ignorecase false

And them I issued the following commands:

$ git mv installer/innosetup/Tribe.iss installer/innosetup/tempname
$ git mv installer/innosetup/tempname installer/innosetup/tribe.iss

Ref:
https://stackoverflow.com/questions/11183788/in-a-git-repository-how-to-properly-rename-a-directory#11183844
https://developers.google.com/style/filenames

Signed-off-by: Thiago Farina <tfransosi@gmail.com>