rename vs project name to lowercase "tribe"
The goal is to have a naming convention to follow when doing C/C++
projects.

The name of solution (.sln) and project (.vcproj) files will be changed to
lowercase as well in the next patches.

This will make it easier to add third party project libraries, like
sqlite or any other project in the future.

Ref: https://developers.google.com/style/filenames

Signed-off-by: Thiago Farina <tfransosi@gmail.com>