rename Tribe.vcproj to all lowsercase tribe.vcproj
The goal is to have a naming convention to following when creating, using
or doing C/C++ projects.

From now on, try to follow the convetion that the project name for a C/C++
project is all lowercase and the file name for the project is also
projectname.vcproj, same for the solution (.sln) file.

This should make the overall structure of the project more consistent and
make it easier to add new projects (our owns or third parties) in the
future.

Ref: https://developers.google.com/style/filenames

Signed-off-by: Thiago Farina <tfransosi@gmail.com>