rename Tribe.sln to all lowercase tribe.sln
The goal is to have a naming convention to follow when creating, using
and/or doing C/C++ projects.

From now on, we will try to follow the convention that the project name
for a C/C++ project is all lowercase. The file name for the project is
also projectname.vcproj, same for the solution (.sln) file.

Overall this should make the structure of directories and files and the
organization of source code more consistent and easier to add new projects
(be it ours or theirs).

Signed-off-by: Thiago Farina <tfransosi@gmail.com>