remove IDS_APP_TITLE resource string
It was wrongly defined.

According to Microsoft's UX Guidelines, the title bar text should be based
on the type of the window.

The types are:
 - Top-level, document-centric program windows: Use a "document name -
   program name" format.
 - Top-level program windows that are not document-centric: Display the
   program name only.
 - Dialog boxes: Display the command, feature, or program from which the
   dialog box came.
 - Wizards: Display the wizard name. Note that the word "wizard" should
   not be included in wizard names.

In our case, we have the second type, so it should display the program
name only, as we were already display by using IDS_APP_NAME rather than
IDS_APP_TITLE.

Ref:
https://learn.microsoft.com/en-us/windows/win32/uxguide/text-ui

Signed-off-by: Thiago Farina <tfransosi@gmail.com>