tribe: calculate the row count of contacts table first
In order to allocate a region of memory large enough for the array to
hold all the contacts, first we need to know how many lines or rows the
table has. Otherwise, we can't tell malloc how big we want this region of
memory to be.

We are allocating an array and populating it, because that is what we plan to
return from 'ab_load_contacts' function. We still populate the linked list
because it's used by the client code. By returning an array, it will be
client's responsability to populate the linked list from it.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>