tribe: initialize database much earlier
Even before calling 'populate_list_view' function. Because we don't want to
continue or even show the appliation window if there is something wrong
with the underlying database.

This is what we already do in the Windows application, so it is good that
the two versions do these steps roughly at the same time.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>