add "ITEM" suffix to CONTACT struct name
We need to do this, because I want to remove the "ROW" suffix from the
other contact struct. Otherwise, it will create a naming conflict and
won't build.

We want to remove the "ROW" suffix from the other struct to make it even
closer to the linux port. We want to reduce the differences between the
two code bases as much as we can.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>