remove "ROW" suffix from CONTACTROW struct name
We are removing the "ROW" suffix because it is repeating itself and not
adding much more valuable information. Contact is already a row in the contacts
table. Another reason for doing this is because in the linux code base,
this structure is called ab_contact_t (not ab_contact_row_t) and we want
to minimize and reduce the differences between them as much as possible.

That change will also allow the introduction of a CONTACTARRAY structure,
rather than a CONTACTROWARRAY, with the former being much more readable
than the later.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>