tribe: change the return value of ab_get_contact_by_id function
Change ab_get_contact_by_id function to return an int, because we
want to return a status (error) code as we already return from
the other functions. Also, that will give us a better way to
verify the success or a failure of the function call.

We want to change this function to select the contact directly
from the database rather than iterating through the linked list,
so we think this change will help with that goal.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>