tribe: fix a memory leak in tribecmd-add.c
We allocate memory for a contact before calling ab_add_contact
function. If it returns an error, we should free the memory
previously allocated for the contact. As we were not doing it,
it was leaking in case of an error.

I did this at commit 3a142a86("sab: check the return value of 'ab_add_contact'"),
but I forgot to port it to tribecmd-add.c.

This commit put these two in sync.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>