tribe: add ab_add_contact_v2 function and use it
The reason for this is that the command line commands (tools or programs) are
single shot operations, they are fire and run. So that this function can
talk directly to the database to perform this operation, and it does not need
to manage the state of the linked list. By avoiding that, it helps with our
goal of making the api as stateless as possible.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>