tribe: add a second version of print_contact_list function
The reason for doing this is that with _db_get_contact_list function
now returning an array, we don't need to fill a linked list with
the contacts just to print them later. We can use the array directly
to print the contact item information.

Another good (and big) reason is that everytime you add something to
the linked list (in alpm_list_t) it allocates memory for a new alpm_list_t.
So we are basically doublying the usage of memory, by letting it fill
the linked list from the array.

Since the list (ls) command is a single shot command, we don't need
to use the linked list, we can just use the array for its purpose.

NOTE: One thing we should do later, that we are not doing right now,
is to free the memory used by the array. That is work for another patch.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>