tribe: free the contacts list array after displaying it
ab_load_contacts_v2 function and thus _db_get_contact_list allocate
memory for the entire array. If it succeeds, we should release
the memory used by it after we are done using it.

Two important notes. One is that it is wrong to iterate through the "array"
and free each indiviual item. Doing that will cause double free error.
free() function seems to know how to free the entire region of memory
allocated previously by the malloc call. The other thing is that, if you
don't assign the pointer variable to NULL after the free call, it will
still have the address of the memory location previously assigned. I
checked that by inspecting it under gdb.

This is a follow up to the work I did at commit 49c2d39 ("tribe: add a
second version of print_contact_list function").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>