tribe: change cmd in "tribecmd" to cli
This is because "tribecmd" is a command line interface (cli) to
the tribe address book. It is not just the "tribe" command (cmd).

This way the program name "tribe" is left for the gui on Linux (GTK)
and Windows (Win32) and "tribecli" for the command line (currently
available only on Linux). Marking clearly the distinction between
the two versions.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>