change the return value of CreateMainWindow function to HWND
When using the verb "Create" for naming a function, it should return a
heap object.

Functions that return BOOL (boolean functions) should be asking true or
false questions, such as IsEmpty(), ShouldClearBuffer(), HasEnoughSpace(),
CanSeek(), etc. By the way, "Is", "Should", "Has" and "Can" are the prefixes
that should be used in boolean functions and they should be in the
affirmative tense.

Refs:
https://git.sr.ht/~sircmpwn/cstyle#variable-and-function-naming
https://docs.unrealengine.com/4.27/en-US/ProductionPipelines/DevelopmentSetup/CodingStandard
https://learn.microsoft.com/en-us/dotnet/standard/design-guidelines/names-of-type-members

Signed-off-by: Thiago Farina <tfransosi@gmail.com>