tribe: rename _db_get_contact_list function to _db_get_contacts
We decided to follow a dotnet design guideline used for naming collection
properties that apply really well in this case here.

It says "DO name collection properties with a plural phrase describing the
items in the collection instead of using a singular phrase followed by "List"
or "Collection"".

That is what we did here, we changed the singlar phrase ("contact") followed by
"list" to a plural phrase ("contacts") which describe the items in the
collection.

We also consulted another guide by John Farrier, which says:
"Plural Naming
Use plural names when referring to collections or multiple instances of items.

Methods that Return Collections: A method that retrieves multiple items should
have a plural name."

Since that is our case, we changed the function name from singular to plural.

Refs:
https://learn.microsoft.com/en-us/dotnet/standard/design-guidelines/names-of-type-members
https://johnfarrier.com/a-pratical-guide-for-naming-things

Signed-off-by: Thiago Farina <tfransosi@gmail.com>