tribe: rename ab_get_contacts function to ab_enum_contacts
We decided to use the verb "enumerate" (enum) rather than "get" because "get"
could be giving the idea of static function that was just returning a
"property", but that is not the case. This is a function that queries the
database to enumerate all known contacts everytime it is called, so it is
much more dynamic than a simple "property" function. And because the meaning
of this verb is "to count off or name one by one; list.", which matches
perfectly with the intent and job of this function.

I have found in the Windows API (Win32) similar functions that show the use
of "enum", as in FaxEnumRoutingMethodsA, FaxEnumJobsA, DhcpV4EnumPolicies. And
some other good examples, like EnumChildWindows, EnumDesktopsA,
EnumDisplayDevicesA, EnumDisplayMonitors, etc.

I also considered the verb "query", and QueryDisplayConfig was an example,
but the use of "enum" seemed much more common. Maybe in the future we can
use "query.

Refs:
https://learn.microsoft.com/en-us/windows/win32/api/winfax/nf-winfax-faxenumroutingmethodsa
https://learn.microsoft.com/en-us/windows/win32/api/winfax/nf-winfax-faxenumjobsa
https://learn.microsoft.com/en-us/windows/win32/api/dhcpsapi/nf-dhcpsapi-dhcpv4enumpolicies
https://learn.microsoft.com/en-us/windows/win32/api/winuser/#functions

Signed-off-by: Thiago Farina <tfransosi@gmail.com>