fix CreateChildrenControls function name
"Child" is a singular noun and "children" is a plural noun even though it
doesn't have an -s. So to fix this poor English in this function name, we
are renaming Children to Child.

The singular also fits better with the A/HC/LC naming pattern that offers
a systematic approach to name entities.

A/HC/LC means Action (A) + High Context (HC) + Low Context (LC), as
desribed in John Farrier's article.

Refs:
https://www.tprteaching.com/children-is-children-are
https://johnfarrier.com/a-practical-guide-for-naming-things

Signed-off-by: Thiago Farina <tfransosi@gmail.com>