fix ab_init function name format
Function names in the Windows code should be in mixed case (CamelCase or
PascalCase) with the first letter of each word capitalized. Hence, it was
renamed to ABInitialize.

While at it, we changed from the abbreviation Init to the full word, the
verb Initialize. Here, we decided to follow the Dotnet naming convetions
and guidelines on using abbreviations and acronyms. They say to NOT use
abbreviations or contractions as part of identifier names. As rule, we say
abbreviations should be avoid in the Windows code base and that divergion
with the Linux code base is fine becasuse this code is not shared between
the Windows platform and the Linux platform, otherwise we would have to
choose between one of the two ways of naming functions and stick with that.

Refs:
https://lean.microsoft.com/en-us/dotnet/standard/design-guidelines/general-naming-convetions#using-abbreviations-and-acronyms

Signed-off-by: Thiago Farina <tfransosi@gmail.com>