rename GetContactList function to ABEnumContacts
This is basically a port of commit 5c6125c ("tribe: rename ab_get_contacts
function to ab_enum_contacts") made in the Linux repository at
https://github.com/tfarina/maple3, to keep the differences between the two
code bases at minimum.

The reason for using the verb "enumerate" or "enum" is because it means
"to count off or name one by one; list". That matches perfectly with the
intent (or job) of this function.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>