tribe: use the verbs "create" and "destroy" with ab_contact_t object
ab_contact_t object is dynamically allocated and thus is memory space
for it is allocated in the heap. For this reason, we should use the verbs
"create" and "destroy" instead of "alloc" and "free", because they are wrappers
around these functions, and do more work than just allocating and freeing memory,
like initializing other variables, etc.

In general, "allocate" or "alloc" and "free" verbs should be left for
low level functions, very low level functions like the ones from standard
library as in malloc and free.

Refs:
https://git.sr.ht/~sircmpwn/cstyle#variable-and-function-naming

Signed-off-by: Thiago Farina <tfransosi@gmail.com>