handle the Delete (Del) shortcut key
Fire a message when the Del key is pressed on the keyboard to delete the
selected item(s) from the list view.

Instead of sending a message (or posting) a message saying the IDM_DELETE
was clicked, we could have added this to the accelerators table, and let
that do the job. We may go this way in the future, for now the code was
already in place and was missing just this piece to route the command.

It worked exactly as if the user clicked in the menu item "Delete" through
the "File" menu.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>