set hr to E_OUTOFMEMORY in LoadContacts function
We need to set it because we are going to change the return type of
LoadContacts function to HRESULT in order to have a better error handling.

This is a preparation step to remove the "return 0" and "return -1", and
have a single "return hr" from the bottom of the function, resulting in a
better control flow and error handling.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>