change the return type of LoadContacts function to HRESULT
By changing the return type to HRESULT and returning only from the very
bottom of the function, it will make room for us to properly release
the memory space allocated for all the objects in this function. Something
that wasn't possible until now, or that would lead to a very ugly code.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>