free the contacts array after using it
We use the contacts array to build a linked list of contact entries. After
using it, we should release the memory allocated for the array to free
the space used by it. Otherwise, it is a waste of memory.

Reducing memory usage and consumption is always good, as it can avoid
unnecessary out of memory errors and unnecessary crashes.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>