tribe: free the contacts array after populating the linked list
Release the memory used by the array after using it, as the entries were copied
over to the linked list and thus it is not necessary to keep the array hanging
in memory.

This was made possible by the work we did at commit a284e64 ("tribe: copy the
contacts array entries over to the linked list"). If we try to do that
without copying the application would crash.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>