tribe: free the contacts linked list
Now that we have a call to ab_fini function, which was added by commit 82b68de
("tribe: add a call to ab_fini function before the application exits"), we have
the opportunity to free the contacts linked list when the application exits.

The macro FREELIST calls alpm_list_free_inner first, which frees the "data"
member from the elements first, then it frees the linked list nodes.

That will allow us to move the maintainance of the linked list into the client
side, liberating the "ab" api from it.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>