rename MAX_LOADSTRING macro to MAX_COLUMN_TEXT_LENGTH
MAX_LOADSTRING macro came with the code generated by the Visual Studio
Project template added by commit 113f56d ("sab: check-in code for a
Windows version"). It is no longer used with LoadString function calls. It
is now specific to CONTACT_ITEM structure, used to determine the number of
characters (or bytes) the columns in the list view can store.

Next, we are going to change the 100 value to something more specific to
the column length.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>