reduce the number of characters allowed in a column
Each character takes 1 byte of memory, so by reducing the number of
characters we can display in each column, we reduce the memory usage.

64 was chosen because it is a power of 2 (2 to the 6 power), which is
better for page alignment, and it was the closest power of 2 that was less
than 100.

This way, the CONTACT_ITEM structure should take 4 bytes for the integer plus
3 x 64 for the other three TCHAR members.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>