tribe: free and null the ab_contact_t object if something goes wrong
Hardening the error handling in ab_contact_create function.

In practice this won't change anything as it is right now, because if malloc
fails and contact is NULL, it will jump to the err label and free will do
nothing and -1 will be returned.

But it is good if we do more stuff and check for errors later on. That way,
we can handle that better and go to the err case and clean memory properly.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>