tribe: free contact object if user cancel the properties dialog
This is a follow up to commit e93d58f ("tribe: create contact object before
opening the properties dialog").

The issue was that if the user opened the properties dialog through "New"
button on the toolbar or throught the "New" menu item in the menubar and
then decided to cancel the dialog, that would leave the memory space reserved
for the contact object laying around without being used anymore and worse
if he did that repeatedly, we would be creating many contact objects in memory
but never releasing the memory if not used.

To fix that issue, we know check in the cancel callback function, if the user
wanted to create a new contact and if we have a pointer to it, if the answers
are true then we release its memory without causing any further trouble.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>