tribe: NULL the contact parameter of ab_contact_destroy function
The call to free() function just releases the memory. The variable contact
still holds the same address value as before the free call.

I checked that under gdb:

(gdb) b ab_contact_destroy
(gdb) p contact
$1 = (ab_contact_t *) 0x8118cb8

Before and after the call to free function, p contact shows the same value.

So, assigning that variable to NULL, will clean up the old address value.

NOTE: That will not affect the variable passed in to ab_contact_destroy
function, that variable will still have the same address value. Yeah! C can
be very weird and misterious in that sense!! Basically, the parameter is a
totally separate variable, even though it is a pointer variable.
HINT: A double pointer parameter might be necessary in this case, to NULL out
the variable that was passed in.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>