tribe: make ab_contact structure pointer variables modifiable
This removes const type qualifier from the pointer variables of the
ab_contact structure.

They should not be constant because we can and modify them at any time
through the set_xxx functions. They should be mutable, because that is
what they are for, they are strings that the user is free to change
whenever they want during the usage of the program.

This fixes the following warnings introduced by the commit bb9870c ("tribe:
duplicate the string value inside the ab_contact_set_xxx functions"):

../src/ab_contact.c: In function ‘ab_contact_set_first_name’:
../src/ab_contact.c:54:10: warning: passing argument 1 of ‘memcpy’ discards ‘const’ qualifier from pointer target type [-Wdiscarded-qualifiers]
   memcpy(contact->fname, fname, len);
             ^
In file included from ../src/ab_contact.c:4:0:
/usr/include/string.h:42:14: note: expected ‘void * restrict’ but argument is of type ‘const char *’
   extern void *memcpy (void *__restrict __dest, const void *__restrict __src,
	                    ^

Refs:
https://learn.microsoft.com/en-us/cpp/c-language/type-qualifiers?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>