remove "Header Files" and "Source Files" filters
The reason for doing this is that I don't want to be opening these virtual
folders all the time (it is time consuming). I prefer to see the source files
right way. Another reason is that I think it is more productive to see the
source file (.c) alongside its complementary header file (.h) right way.

This flatten project structure looks much better and cleaner in my opinion.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>