add the initial code to create a new contact
This part of the code is very complex and will get more complicated as
more code is added. That is why I'm splitting it in many small parts.

Eventually, when all the code is put in place, it will insert a contact
into the sqlite database and display it on the list view.

I had to use the SetWindowLongPtr/GetWindowLongPtr trick, because I'm
sending the pointer to the property sheet through its lParam and that
comes with WM_INITDIALOG message. When the other messages get by, the only
way to retrieve the pointer to the PROPDLGINFO would be if it was set when
WM_INITDIALOG was fired, and we will need it when handling the
PSN_KILLACTIVE message. I used GWLP_USERDATA because it is a user
data (property dialog information) associated with this property sheet. I
could have used DWLP_USER too though.

Anyway, this is just the start point, and many more things need to be done
to get it to the end and towards our goal of storing the contact
information into the database.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>