embed tribe.exe.manifest file into the final binary
From the very beginning we wanted to have it this way, but never got to
the point of setting things the right way, until now.

By removing the manifest properties and letting the default values from
Visual Studio settings kick in, we fixed the logic.

We want it to generate a manifest and embed our manifest file into the
final assembly. I realized this after looking and reading the
source code tree of the first few commits of Chromium for Windows.

Since this change is small and very contained, if any problem of getting
the Visual Styles of Windows XP comes back, we can revert this change to
get the old behavior.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>