set up the contact object to manipulate its properties
This is the object we are going to use to create a new contact for
manipulating all its properties.

This is the second part of a series ofpatches to get a contact object
inserted into the sqlite database. More things need to be done to get
there.

This is a follow up to commit e0a97c ("add the initial code to create a
new contact").

WARN: There is a memory issue that will keep increasing as the user keeps
clicking on the "New" button on the toolbar or the menu, because we are
not releasing it when we are done manipulating it. We will have to look at
this and fix it later on.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>