tribe: move sqlite3_finalize call into the cleanup area
This is to unify the code path for deleting the prepared statement.
By having it under the out label, we make sure it is called from
any goto statement.

I also confirmed that calling sqlite3_finalize() on a NULL pointer is a harmless no-op.
It does not crash the application, so checking the statement before calling it is not
necessary and thus I removed this check as well.

Refs:
https://www.sqlite.org/c3ref/finalize.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>