tribe: initialize all local variables in _db_insert_contact function
This is to make sure they have some sane initial value.

We don't need to jump to the cleanup area if sqlite3_prepare_v2() function
returns an error because the documentation says insert_stmt will be set to
NULL and thus calling sqlite3_finalize would be pointless as it won't
do anything.

Refs:
https://www.sqlite.org/c3ref/prepare.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>