split version number in individual macros
The great benefit of doing in this way is that it allows changing a single
line for a specific version number, be it major, minor, patch or revision.
Making git blame much more helpful when blaming the file or seeing its
commit history. That makes changes much more easily identifiable.

We decided to use the version format of major.minor.patch.revision from
https://semver.org for its simplicity.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>