build ProductVersion string from version number macros
The great benefit of doing in this way is that now changing one of the
version number macros, say for example, changing MINOR to 2, it won't be
necessary to make changes to VI_PRODUCTVERSION and VI_STR_PRODUCTVERSION
like before. They are now kept in sync from now on and this avoid
potential mistakes like updating one but forgetting to update the other
and vice-versa, as it could occur with the previous version.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>