insert contact object into the sqlite database
With this patch the user can now enter the first name, the last name and
email in the Properties dialog, click OK, and the data will get saved into
the contacts table.

This is the third part of series of patches to get a contact object
inserted into the sqlite database.

This is a follow up to commit c20d64f ("set up contact object to
manipulate its properties").

To confirm that the contact information got added to the contacts table, I
opened sqlite3 shell (sqlite3.exe) on commnand line and ran the following
commands:
sqlite> .open adbd.sqlite3
sqlite> select * from contacts;
1|John|Smith|smith@mail.com
2|John|Doe|john@hotmail.com

NOTE/TODO: After this, we have to work on to get this new information
displayed in the list view.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>