tribe: add an errcode variable to _db_delete_contact function and use it
The issue is that with the sqlite3_finalize call and its error handling.
If it returns an error and we enter the if statement, then we print
an error message, but we end up returning a success status code (0) with
the "return 0" statement right after the sqlite3_changes call, which is
wrong. By setting the errcode variable to -1 and removing the "return 0"
statement we fix that.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>