tribe: fix pointer-sign errors in _db_enum_contacts and ab_get_contact_by_id functions
This fixes the following kind of error when enabling -Werror for source code
under src directory:
../src/ab.c: In function ‘_db_enum_contacts’:
../src/ab.c:240:35: error: pointer targets in passing argument 1 of ‘xstrdup’ differ in signedness [-Werror=pointer-sign]
       contacts[i].fname = xstrdup(sqlite3_column_text(select_stmt, 1));

I'm omitting all the errors reported, because the output is big,
and it is more of the same error above.

The issue was with the sign of the char data type. sqlite3_column_text
returns unsigned char const *, while xstrdup takes char const *. So,
in order to fix the warning (turned into error by -Werror), was necessary
to cast it.

Refs:
https://sqlite-users.sqlite.narkive.com/nJ6r4F4f/strcpy-and-sqlite3-column-text

Signed-off-by: Thiago Farina <tfransosi@gmail.com>