tribe: add "g_" prefix to progname variable
A variable declared with the "static" specifier has a global lifetime,
and thus exists and has a value throughout the execution of the program.

Variables declared outside of all function definitions appear at the
"external level".

So, in order to clearly differentiate this kind of variable from
automatic variables or variables with local lifetimes (declared inside
functions), we are adding the prefix "g_" (global) to progname variable.

Refs:
https://en.wikipedia.org/wiki/Hungarian_notation
https://learn.microsoft.com/en-us/cpp/c-language/c-storage-classes?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>