tribe: remove cmake's project command from sqlite build file
Why? Because we build it as part of the tribe project and don't want it
to set the name of the project as sqlite and to mess with the PROJECT_NAME
variable. The top-level CMakeLists.txt is responsible for that, and the only
one that should do that.

If it was built in a separate directory, as in
https://github.com/azadkuh/sqlite-amalgamation/, that would be another story.

Refs:
https://cmake.org/cmake/help/latest/command/project.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>