tribe: define a desktopdir variable and use it
This is to mirror the automake code in Makefile.am that does this:

desktopdir = $(datadir)/applications
desktop_DATA = $(desktop_in_files:.desktop.in=.desktop)

We also created a DATADIR variable and defined it the same way automake does.
It will serve for defining a themedir variable for the icons as well.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>