tribe: put config.h.cmake in the top-level directory
Although this file was created by us instead of by a program, it is analogous
to the template file config.h.in created by autoheader.

By invoking autoheader in the top-level directory, it scans the configure.ac,
and generates the template file in the top-level directory.

So, since this is a header to keep compatibility with autotools, let's have
config.h.cmake in the top-level directory as well.

The input (config.h.cmake) is a relative path, so according to cmake's documentation
it is evaluated with respect to the current source directory.

Refs:
http://web.mit.edu/gnu/doc/html/autoconf_3.html
https://www.gnu.org/software/autoconf/manual/autoconf-2.67/html_node/autoheader-Invocation.html
https://www.gnu.org/software/autoconf/manual/autoconf-2.67/html_node/Configuration-Headers.html#Configuration-Headers - AC_CONFIG_HEADERS
https://cmake.org/cmake/help/v3.0/command/configure_file.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>