add a vcproj file for sqlite library
Why? Because this allow us to have the sqlite source code in their own
project and then build it as a separate static library (sqlite.lib).
Segregating the code of a external project from the build of the main
code.

Currently only the project file was added to the solution file and
verified to build in Debug and Release configurations with the default
options created from the template installed by Visual Studio. The tribe
project does not depend on it yet and still builds sqlite3.c file.

Some of the setting of these options need tweaking and others need to be
added.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>