disable RuntimeTypeInfo option (/GR-)
RuntimeTypeInfo (RTT) is a mechanism to determine the type of an object
during program execution added to C++ language. It is for code that uses
dynamic_cast Operator or typeid. This option increases the size of .rdata
section. Since our code is written in C we can disable this feature.

For more information, see the reference.

Refs:
https://lean.microsoft.com/en-us/cpp/build/reference/gr-enable-run-time-type-information?view=msvc-140

Signed-off-by: Thiago Farina <tfransosi@gmail.com>