tribe: move the check for required packages to the top-level CMakeLists.txt file
This kind of check in the Autotools build system is usually done in a top-level
configure.ac file in the top-level directory.

In order to make our cmake build files more compatible with autotools, let's
have this check in the top-level CMakeLists.txt file as well.

Tested with the following commands:

$ rm -r out
$ mkdir out
$ cd out
$ cmake -GNinja -DCMAKE_BUILD_TYPE=Debug ..
$ cd ..
$ ninja -C out
$ ./out/bin/tribe

Signed-off-by: Thiago Farina <tfransosi@gmail.com>