tribe: pass GTK_LDFLAGS to target_link_libraries
GTK_LDFLAGS seems more correct than GTK_LIBRARIES for
the use of pkg_check_modules, because the documentation
says the value of GTK_LDFLAGS is all required linker
flags while for GTK_LIBRARIES is only the libraries (without the '-l').

Where with FindGTK2, the value for GTK2_LIBRARIES is all libraries.

Both ways lead to the same command like though, which was:

[17/17] : && /usr/bin/gcc -Wno-unused-parameter -Werror -Wextra -Wall -std=c99  -g -rdynamic src/ui/gtk/CMakeFiles/tribe.dir/about_dialog.c.o src/ui/gtk/CMakeFiles/tribe.dir/contact_editor.c.o src/ui/gtk/CMakeFiles/tribe.dir/main.c.o src/ui/gtk/CMakeFiles/tribe.dir/window.c.o -o bin/tribe  src/libab.a  src/third_party/sqlite/libsqlite.a  -lgtk-x11-2.0  -lgdk-x11-2.0  -lpangocairo-1.0  -latk-1.0  -lcairo  -lgdk_pixbuf-2.0  -lgio-2.0  -lpangoft2-1.0  -lpango-1.0  -lgobject-2.0  -lglib-2.0  -lfontconfig  -lfreetype  -pthread  -ldl && :

Refs:
https://cmake.org/cmake/help/latest/module/FindPkgConfig.html
https://cmake.org/cmake/help/latest/module/FindGTK2.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>