free the contact object after using it
This releases the memory space allocated for the CONTACT object after the
Properties dialog is closed. In any case, be it when just creating a new
contact or showing its properties, the contact memory should be released,
because the object is created only to be manipulated at the Properties
dialog. The data information that appears in the list view is transfered
(copied) from this contact object onto the CONTACT_ITEM, so only the
CONTACT_ITEM object needs to stay around (alive).

Tested creating a new contact and filling the information and pressing the OK
button and also tested creating a new contact but canceling the dialog. In
both cases the program didn't crash.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>