check the return value of PropertySheet function
This checks the return value of PropertySheet() function to decide if we
should continue and add a new item to the list view or abort.

For modal property sheets (like ours), the return value is as follows:

>=1 Changes were saved by the user.
0   No changes were saved by the user.
-1  An error occurred.

Zero is returned when the user clicks the Cancel button or closes the
property sheet dialog either through the Close (x) button or with Esc.
>=1 is returned when the user clicks the OK button.

Refs:
https://learn.microsoft.com/en-us/windows/win32/api/prsht/nf-prsht-propertysheeta

Signed-off-by: Thiago Farina <tfransosi@gmail.com>