tribe: fix the indentation mess in _remove_selection function
There was a indentation mess in _remove_selection function because we mixed
different types of formatting indentation.

The GNU style is pretty ugly, the K&R is fine though, but from what I have
seem the one that is most readable is the style used on Windows API (which
seems to be the Allman style), which puts the curly braces in their own line
and indents 4 spaces (we are keeping two spaces because the rest of the
code is indented with 2 spaces).

See some examples below:

if (expression)
{
  statement
}

for (init-expressionopt; cond-expressionopt; loop-expressionopt)
{
  statement
}

while (expresion)
{
}

Refs:
https://en.wikipedia.org/wiki/Indentation_style#Allman_style

Signed-off-by: Thiago Farina <tfransosi@gmail.com>