tribe: remove contact entry from database first
This database operation is more important that removing the item from the list
view control. The main goal of this function is to remove the contact entry
from the database, so this operation should be done first. Later, the UI
should be updated based on the results of that operation.

Tested deleting some entries obviously.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>