remove project files for Visual C++ 6.0
I don't have time and resources (money) to make sure everything works when
building with Visual C++ 6.0 IDE. It is very outdated and some things do
not work as expected. So I prefer to put all my effort on making sure
building on Visual Studio 2005 works as smoothly as possible.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>