define OutputFile property using OutDir variable
$(OutDir) resolves to the value of the Output Directory, which in our case
for now is either src\Debug or src\Release (but we might change that in the
future) depending on the configuration selected.

This property change be changed on Visual Studio 2005 selecting the
project on Solution Explorer -> Properties -> Linker -> General.

Tested cleaning up both Debug and Release configurations and doing a full
build in both configurations.

Refs:
https://learn.microsoft.com/en-us/cpp/build/reference/common-macros-for-build-commands-and-properties?view=msvc-170&redirectedfrom=MSDN
https://learn.microsoft.com/en-us/cpp/build/reference/out-output-file-name?view=msvc-140

Signed-off-by: Thiago Farina <tfransosi@gmail.com>