tribe: stop removing items from contacts linked list
The old way required us to do that because get_contact_by_id
was using the linked list. Now that get_contact_by_id does not
use the linked list anymore, we don't need to remove an item
from the linked list everytime a contact is deleted. All
the memory used by it will be released when ab_fini is called
and FREELIST is used.

Also, there is a performance issue related to this. Everytime
we want to remove an item from the linked list it has to iterate
through the list comparing the ids to find the exact one to remove.
Imagine if we select more than one to be removed, it will loop as
many times as items selected. Of course, the list will shrink,
but doing less work is better in this case.

This will help us merge ab_delete_contact_v2 with ab_delete_contact,
now that they do exactly the same thing.

This also helps with our goal of making ab api stateless as possible.
That is going to happen when all control and maintainance of contacts
linked list is moved into the client side.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>