add item to list view
When the user selects "New Contact" on the menu bar or "New" on the
toolbar, fills the details in the Properties dialog and presses the "OK"
button at the end, this will now add a contact item (represented by
CONTAT_ITEM struct) to the list view.

From the dwResultAction member of PRINTDLGEXA struct I borrowed the idea
for the ulResultAction [1].

This is a big milestone, because now that we are inserting entries in the
database and displaying them on list view, we can now move on to select and
retrieve these entries from the contacts table instead of building an
array with fake entries.

Refs:
[1] -
https://learn.microsoft.com/en-us/windows/win32/api/commdlg/ns-commdlg-printdlgexa

Signed-off-by: Thiago Farina <tfransosi@gmail.com>