move sqlite3 files one level up
This moves the sqlite3 soure files out of the amalgamation directory,
moving them one level up.

The problem we are trying to fix here is to have the word "amalgamation"
in the include path. It is uncommon to have a directory named like this
and it is hard to type. Also, the fact that sqlite3.c is a concatenation
of over 100 separate source files is not an information relevant for
clients including this code. So let's hide this fact.

This is a port of commit a5a97cd ("tribe: move sqlite3 files up one
level") made at https://github.com/tfarina/isengard/.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>