tribe: stop checking the number of columns returned
Checking the number of columns returned is not helping
much here in this case as we are only interested in
the how many of rows the table has. If column_count
returned zero or more than one, the code was not going
to be compromised, as we were going to retrieve the value
stored on the first column anyway, which for our query
is the number of rows in the table.

Removing this check has two benefits. First, it makes
the code more generic, as the check was too narrow to
that specific COUNT sql query. Second, if the check
didn't pass, we would be returning in the middle of the
function, and that would prevent a call to finalize the
prepared statement, which is no good. To prevent that,
we would have to use a goto statement, but I don't want
move to that right now.

So, for now, the simpler approach of removing the check
will suffice.

I tested with the table empty and with some records. Both
worked fine.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>