tribe: ignore sqlite3_finalize errors on purpose
Doing work on goto area has some big problems. Setting
errcode here is wrong because if an error occured and it was already set there
before jumping to this area, then we are basically discarding, disregarding
and ignoring that by setting it to a new value based on the result of
sqlite3_finalize call. Throughing an error message from goto area might be
okay, but if we came here because of out of memory error, then trying to
write out a message might not be the best thing to do. Instead, what we
should try to do in this area is to free and release resources as much and as
soon as possible and do as minimum work as possible. Setting the pointer to
sqlite statement (sqlite3_stmt *) to NULL also might be okay, but has no
practical benefits as the variable is local and will be initialized to NULL
when the function is called again.

The Linux Kernel coding style [1] gives some explanation for why using
goto statements:

"
The goto statement comes in handy when a function exits from multiple
locations and some common work such as cleanup has to be done.  If there is no
cleanup needed then just return directly."

And some reasons for using it too:

"
The rationale for using gotos is:

- unconditional statements are easier to understand and follow
- nesting is reduced
- errors by not updating individual exit points when making
  modifications are prevented
- saves the compiler work to optimize redundant code away ;)
"

More information about the style we use for goto statements can be found
at the cprogramming [2] article.

I have tested this manually, by running the standalone commands
(tribels, tribeadd, tribemod, tribedel), through tribecli and also
through the gtk ui with tribe.

Refs:
[1] - https://kernel.org/doc/Documentation/process/coding-style.rst
[2] - https://www.cprogramming.com/tutorial/goto.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>