tribe: cleanup sqlite3_finalize code in _db_enum_contacts function
As a result of moving it into the cleanup area (goto's labeled statement),
it does not overwrite errcode value anymore if it was set previously.

As the cleanup area is meant for freeing and releasing resources only, it stops
checking the return value of sqlite3_finalize and issuing an error message if
the check fails.

This is a follow up to the work done at commit aa7ec86 ("tribe:
ignore sqlite3_finalize errors on purpose").

Tested running the following:

>tribe

>tribels

>tribecli list
>tribecli ls

Signed-off-by: Thiago Farina <tfransosi@gmail.com>