tribe: fix memory leak in cmd_delete.c and del.c
This fixes a memory leak caused by calling ab_get_contact_by_id
function but never freeing the contact object returned.
ab_get_contact_by_id calls ab_contact_create, which in turn calls
malloc to create the contact object for the id the user solicited.
When ab_get_contact_by_id returns, if it succeeded, then memory was
allocated for the contact type and we need to release it when
we are not using it anymore. And that is what we do now in the clean
up area.

Tested with the following:

>tribecli delete id-number
>tribecli del id-number
>tribedel id-number

And also inspected in gdb to make sure everything worked properly.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>