tribe: close the database when an error occurs in cmd_modify.c and mod.c
This fixes the error handling in cmd_modify.c and mod.c,
by not returning emidiately when an error occurs after
the call to ab_init.

After ab_init is called, we need to make sure ab_fini
is called in order to close the database properly to
avoid any corruption and loss of data.

We don't have a memory leak related to ab_get_contact_by_id
because we already call ab_contact_destroy in both cases (
success and failure) to release the memory allocated for
the contact object, and that code is preserved in the cleanup
area.

Tested running the following:

>tribecli modify id-number ...
>tribecli mod id-number ...
>tribemod id-number
>echo $? # to check the return value

I usually try with -1, 0, and other non-existant ids, to
make sure at least one error case is triggered.

I also went line by line under gdb, stepping trough, to make
sure everything is behaving properly and working as before.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>