tribe: close the database when an error occurs in tribe executable
As ab_init may fail for various reasons (it might not be able
to open the database, the integrity check may fail, etc)
we should not return emidiately without calling ab_fini.

If it fails, we should call ab_fini to make sure the database
is closed properly to avoid any corruption and loss of data.

This fixes this error handling by using a goto statement,
as it keeps the code clean and easy to read.

Also note that we were calling ab_fini before quitting
the gtk main message loop. By moving this call into the
clean up area under the "out" label in the main function,
it also works after quitting the message loop, as it will
go through there as part of the normal follow.

Tested running the following:

>tribe

And under gdb with the following to make sure it gets called after
the message loop is quit:

>gdb ./out/bin/tribe
b main.c:38

This is a follow up to commit d9ceb1b ("tribe: close the database when an
error occurs in cmd_modify.c and mod.c").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>