tribe: check sqlite3_step for error in _db_get_row_count function
This changes the check of the value returned by sqlite3_step function from
success to failure, as we check for error everywhere else. If it succeeds,
sqlite3_step will return SQLITE_ROW. This allow us to print an error message,
set a status code for the error and jump to the cleanup area.

This will also allow us to get rid of the local variable "row_count" used
to store the value of sqlite3_column_int function temporarily, as it is now
more self-evident then it was before that it is not necessary to have it.

Tested running the following:

>tribe
>tribecli list
>tribecli ls
>tribels

Signed-off-by: Thiago Farina <tfransosi@gmail.com>