use negative ids for the fake entries
I'm changing the ids to negative values for the fake entries to avoid
clashing with the real entries in the database. Sqlite will create entries
with positive ids, starting with 1, 2, 3, etc. So when we start loading
the entries from the database they will override the fake ones in the
transition period. In order to avoid that, let's use these negative ids. I
didn't start with -1, because we may use -1 as an error indicator.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>