tribe: fix potential memory leak in _db_enum_contacts function
We allocate memory for the contacts array before preparing the sql statement.
If it succeeds and then the sqlite3_prepare_v2 fails, we have a memory leak,
because we don't free the memory allocated for the array in the error path
of the sqlite3_prepare_v2 function. Jumping to the cleanup area should fix
this leak.

Tested running the following:
>tribe
>tribecli list
>tribecli ls
>tribels

Signed-off-by: Thiago Farina <tfransosi@gmail.com>