c: use NULL macro when returning an error in xstrdup function
This is to make the intent of the code more clear. It is easier
when using "grep" too.

On my system, the NULL macro is defined for C in [1] as:

((void *)0)

We can assign it to any pointer variable because it has type void *,
so there is no problem if the return value is char *.

Refs:
[1] - /usr/include/i386-linux-gnu/unicode/utypes.h and /usr/include/libio.h
https://www.gnu.org/software/libc/manual/html_node/Null-Pointer-Constant.html
https://learn.microsoft.com/en-us/cpp/c-language/null-macro?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>