avoid using redundant != or == in conditions
Currently, the code in ABEnumContacts() function does NULL pointer
verifications using "ptr != NULL" and "ptr == NULL" and even using Yoda
conditions.

But since the expression can be of a pointer type and the body of the if
statement (same for while and for statements) is executed if the value of
the expression is nonzero, this can be simplified to if (ptr) and if
(!ptr).

This has some benefits. It makes the code less verbose. Avoids an
accidental assignment if some misses one equal sign ("="), the use of Yoda
conditions was guarding against this though. And thus helps the code
become consistent and conventional.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>