introduce a clean up area in ABEnumContacts()
Currently, when the allocation of space in memory for the array fails we
return immediately with E_OUTOFMEMORY.

That is okay for now, as no other allocation happens before.

But as the execution of the code proceeds, we need a way to free its
memory if a failure occur in another location, as it may happen in the
string allocations in the for statement.

So, in order to have a clean up area for releasing resources, we
introduced an "err" label.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>