jump to clean up area when memory alloation for strings fails
Currently, we allocate memory space for the strings in ABEnumContacts()
function and return immediately with E_OUTOFMEMORY if LocalAlloc() fails.

But that is bad, because *before* these allocations, memory space has
already being allocated for the array and thus needs to be freed if any
allocation failure happens *after* it.

So, let's jump to the clean up area to have the opportunity to release all
the resources that need to be free.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>