fix parameter naming in ABEnumContacts()
Currently, pnCount parameter in ABEnumContacts() function is poorly named.

In Hungarian notation [1], "n" prefix may symbolize "integer" in Systems
notation or "count" in Apps notation. So, if it is interpreted in Apps
notation the variable name reads as "pointer to count of Count". Too
generic, not great. If it is interpreted in Systems notation, then it
reads as "pointer to integer of Count". Pretty bad too.

There are other Hungarian prefixes [2] we can use here to improve this
variable name. One is "ul" for the data type of this variable, which
stands for "Unsigned long". The other is "c", for "count of".

With these prefixes, we can rename it to "pulcContacts", and have it read
as "pointer to a unsigned long count of Contacts".

Hopefully, this is a better name.

Refs:
[1] - https://en.wikipedia.org/wiki/Hungarian_notation
[2] -
https://learn.microsoft.com/en-us/windows/win32/stg/coding-style-conventions

Signed-off-by: Thiago Farina <tfransosi@gmail.com>