fix variable naming in ABEnumContacts()
Currently, there are some issues with the name of lpContactList variable
in ABEnumContacts() function. First, it is using a singular noun
"Contact", but it is referring to a collection (an array) of contacts. The
second is that "Contact" is followed by the suffix "List", which might be
okay in this case as "Contact" is singular and "List" indicates it is a
collection, but "List" might be interpreted as referring to a "Linked
List", which is not the case. Here, lpContactList represents a dynamic
allocated array, not a linked list. So, the choice of the name for this
variable is not good.

In Hungarian notation [1] there is a prefix we can use for dynamic arrays.
It is the "rg" prefix and it reads as "range of" or "array of".

With this prefix and using plural, we can rename it to "lprgContacts", and
have it read as "(long) pointer to a range of contacts" or "(long) pointer
to an array of contacts". Hopefully, this is better than what we had
before.

Arguably, I could have use an "Array" suffix and have it as
"lpContactsArray". That is an alternative that we may use for other
variables in the future. For now, let's go with the Hungarian "rg" prefix.

For the future, I will list the rules I encountered online and used for
plural naming below:

"6. Use Plural for Collections:
When naming variables that represent collections (arrays, lists, etc.),
use plural names to indicate that it's not a single element." [2]

"Plural Naming
Use plural names when referring to collections or multiple instances of
items." [3]

.NET
"DO name collection properties with a plural phrase describing the items
in the collection instead of using a singular phrase followed by "List" or
"Collection"." [4]

Refs:
[1] - https://en.wikipedia.org/wiki/Hungarian_notation
[2] - https://dotnetustad.com/c-programming/c-variable-naming-convention
[3] - https://johnfarrier.com/a-pratical-guide-for-naming-things/
[4] -
https://learn.microsoft.com/en-us/dotnet/standard/design-guidelines/names-of-type-members

Signed-off-by: Thiago Farina <tfransosi@gmail.com>