fix memory leak in ABEnumContacts()
First, we allocate space in memory for the contacts array. Then, we
allocate memory for the strings in the for loop. If the allocation of
memory space for the contacts array succeeds but a later one fails in the
loop, then a memory leak occurs, because it is not released (freed) in the
error path.

So, let's fix this memory leak, releasing the memory if an error occurred.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>