introduce a helper function to duplicate strings
We repeat the same code pattern do dupliate a string three times in the
for loop of ABEnumContacts() function.

Imagine if there was many more strings to duplicate for the CONTACT
structure. We would have a lot of code duplication.

So, let's extract this code pattern in a helper function called XStrDup()
and clean up the code in the for loop.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>