tribe: fix variable naming in _db_get_row_count()
Currently, there is a naming issue with the parameter "p_row_count"
in _db_get_row_count() function that makes it look different than the
rest of the code in this file.

The problem is that it is using a Hungarian notation [1], while the rest
of the file does not use it for naming its variables.

The Hungarian prefix "p" stands for for "pointer to" and thus "p_row_count"
reads as "pointer to row count". In this case, it is almost fine, as it is
not vague or too generic and at least it is using the snake_case, which is
the predominant style for naming variables in this file.

So, in order to make it look like the rest of the file and keep the style
consistent, let's remove the Hungarian prefix and leave it for code on the
Windows side.

Refs:
[1] - https://en.wikipedia.org/wiki/Hungarian_notation

Signed-off-by: Thiago Farina <tfransosi@gmail.com>