fix the order in which ABInitialize() is called
In commit c51188cf6aca ("WunderPeople: we should open and create the
sqlite database very early on") we moved the call to ABInitialize() from
the handling of the WM_CREATE constant to right before the call to
LoadAccelerators() functions, but after calling CreateMainWindow()
function. And that was a mistake, because Windows sent WM_CREATE message
before, and all code in the case statement was executed before
ABInitialize() could even be called.

I noticed this code flow mistake while porting the
_db_get_contacts_row_count() function from the Linux code base and saw the
list view with empty result.

To fix this, let's call ABInitialize() before calling CreateMainWindow().
So, functions depending on sqlite being opened will have a valid "hdb"
pointer to work on.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>